requirements.txt
# pip install -r requirements.txt

requests
pandas
openpyxl
pytesseract
Pillow
